+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## Vertion 9.16.0
+_26 March 2018_
+
+### Minor features
+* Added Participant/Age chart to /charts page
+* Added Participant/Education chart to /charts page
+* Added Participant/Gender chart to /charts page
+* Added Participant/Professional Experience chart to /charts page
+* Added Participant/Product Experience chart to /charts page
+* Added Participant/Computer Experience chart to /charts page
+
+---
+
 ## Version  9.15.0
 _12 March 2018_